% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPairsBoot.R
\name{calc_boot_out}
\alias{calc_boot_out}
\title{Calculations from boot return}
\usage{
calc_boot_out(boot_list_return, type_results, nBoot, startSeed, blockLength)
}
\arguments{
\item{boot_list_return}{The object that is returned from \code{run_bootstraps}.}

\item{type_results}{data frame returned from either \code{\link[EGRET]{runGroups}}
or \code{\link[EGRET]{runPairs}} depending on context.}

\item{nBoot}{the maximum number of bootstrap replicates to be used, typically 100}

\item{startSeed}{sets the random seed value. This is used to make repeatable output.}

\item{blockLength}{integer size of subset, expressed in days.  200 days has been found to be a good choice.}
}
\description{
Calculations from boot return
}
\keyword{internal}
