% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ciData.R
\docType{data}
\name{Choptank_caseSetUp}
\alias{Choptank_caseSetUp}
\title{Example caseSetUp}
\description{
Example data representing data from the Choptank River at Greensboro, MD,  USGS data
Data is a a data frame used to set up a trends analysis.
}
\keyword{data}
\keyword{quality}
\keyword{water}

