% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ciData.R
\docType{data}
\name{Choptank_CIAnnualResults}
\alias{Choptank_CIAnnualResults}
\title{Example CIAnnualResults}
\description{
Example data representing data from the Choptank River at Greensboro, MD,  USGS data
Data is a a data frame returned from \code{ciCalculations}.
}
\keyword{data}
\keyword{quality}
\keyword{water}

