% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ciData.R
\docType{data}
\name{Choptank_repAnnual}
\alias{Choptank_repAnnual}
\title{Example repAnnualResults}
\description{
Example data representing data from the Choptank River at Greensboro, MD,  USGS data
Data is a named list of the Daily, Sample, INFO dataframes, and the surface matrix.
}
\keyword{data}
\keyword{quality}
\keyword{water}

