% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCIs.R
\name{ciCalculations}
\alias{ciCalculations}
\title{ciCalculations}
\usage{
ciCalculations(eList, ...)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes. Created from the EGRET package, after running \code{\link[EGRET]{modelEstimation}}.}

\item{\dots}{optionally include nBoot, blockLength, or widthCI}
}
\description{
Interactive function to calculate WRTDS confidence bands
}
\examples{
library(EGRET)
eList <- Choptank_eList
\dontrun{
CIAnnualResults <- ciCalculations(eList)
}
}
