% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wBTCode.R
\docType{package}
\name{EGRETci-package}
\alias{EGRETci}
\alias{EGRETci-package}
\title{EGRETci package for bootstrap hypothesis tests and confidence interval analysis for WRTDS (Weighted 
Regressions on Time, Discharge, and Season) statistical models. This 
package is designed to be used in conjunction with the EGRET package, 
which estimates and describes WRTDS models.}
\description{
\tabular{ll}{
Package: \tab EGRETci\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
\url{https://www.usgs.gov/information-policies-and-instructions/copyrights-and-credits}\cr
LazyLoad: \tab yes\cr
}
Collection of functions to evaluate uncertainty of results from water quality analysis using 
the Weighted Regressions on Time Discharge and Season (WRTDS) method. This package is an add-on 
to the EGRET package that performs the WRTDS analysis.
}
\references{
Hirsch, R.M., and De Cicco, L.A., 2015, User guide to Exploration and Graphics for RivEr Trends 
(EGRET) and dataRetrieval: R packages for hydrologic data: U.S. Geological Survey Techniques and Methods book 4, 
chap. A10, 94 p., \doi{10.3133/tm4A10}

Hirsch, R.M., Archfield, S.A., and De Cicco, L.A., 2015, 
A bootstrap method for estimating uncertainty of water quality trends.  
Environmental Modelling & Software, 73, 148-166. 
\url{https://www.sciencedirect.com/science/article/pii/S1364815215300220}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://code.usgs.gov/water/EGRETci}
  \item Report bugs at \url{https://github.com/DOI-USGS/EGRETci/issues}
}

}
\author{
Robert M. Hirsch \email{rhirsch@usgs.gov}, Laura De Cicco \email{ldecicco@usgs.gov}
}
\keyword{internal}
