% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{estSliceSurfacesSimpleAlt}
\alias{estSliceSurfacesSimpleAlt}
\title{surface slice}
\usage{
estSliceSurfacesSimpleAlt(eList, year)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes. Created from the EGRET package, after running \code{\link[EGRET]{modelEstimation}}.}

\item{year}{integer year}
}
\value{
surfaces matrix
}
\description{
Creates surface slice for one year.
}
\examples{
library(EGRET)
eList <- Choptank_eList
eList <- setForBoot(eList)
surfaces <- estSliceSurfacesSimpleAlt(eList, 1990)
}
\keyword{WRTDS}
\keyword{flow}

