% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ciData.R
\docType{data}
\name{Choptank_eBoot}
\alias{Choptank_CIAnnualResults}
\alias{Choptank_caseSetUp}
\alias{Choptank_eBoot}
\alias{Choptank_repAnnual}
\title{Example eBoot}
\description{
Example data representing data from the Choptank River at Greensboro, MD,  USGS data
Data is a named list of the Daily, Sample, INFO dataframes, and xConc, and xFlux vectors.
}
\keyword{data}
\keyword{quality}
\keyword{water}

