% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_bootstrapping.R
\name{genDailyBoot}
\alias{genDailyBoot}
\title{WRTDSKalman Bootstrapping}
\usage{
genDailyBoot(
  eList,
  nBoot = 10,
  nKalman = 10,
  rho = 0.9,
  setSeed = NA,
  jitterOn = FALSE,
  V = 0.2
)
}
\arguments{
\item{eList}{is the data with a fitted model already done. Note that the eList$Sample 
may have multiple values on a given day and it can also have censored values.}

\item{nBoot}{number of times the bootstrap resampling and model estimating is done}

\item{nKalman}{number of different realizations for each re-estimated model}

\item{rho}{numeric the lag one autocorrelation. Default is 0.9.}

\item{setSeed}{value. Defaults is \code{NA}, which will not specify a randomized seed.
This can be used to make repeatable output.}

\item{jitterOn}{logical. If \code{TRUE}, the code will "jitter" the}

\item{V}{a multiplier for the sd of the LogQ jitter. for example V = 0.02,
means that the sd of the LnQ jitter is 0.02*sdLQ}
}
\description{
Function to get bootstrapping results from WRTDS_K
}
\examples{

eList <- EGRET::Choptank_eList
dailyBoot <- genDailyBoot(eList, nBoot = 2, nKalman = 2)


}
