% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_bootstrapping.R
\name{makeMonths}
\alias{makeMonths}
\title{make Months}
\usage{
makeMonths(dailyBootOut, eList)
}
\arguments{
\item{dailyBootOut}{data frame returned from \code{\link{genDailyBoot}}}

\item{eList}{is the data with a fitted model already done. Note that the eList$Sample 
may have multiple values on a given day and it can also have censored values.}
}
\description{
Month statistics
}
\examples{
eList <- EGRET::Choptank_eList
dailyBoot <- genDailyBoot(eList, nBoot = 2, nKalman = 2)
monthPcts <- makeMonths(dailyBoot, eList)
}
