% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{makeTwoYearsResults}
\alias{makeTwoYearsResults}
\title{makeTwoYearsResults}
\usage{
makeTwoYearsResults(eList, year1, year2)
}
\arguments{
\item{eList}{named list}

\item{year1}{integer}

\item{year2}{integer}
}
\value{
surfaces matrix
}
\description{
makeTwoYearsResults
}
\examples{
library(EGRET)
eList <- Choptank_eList

twoResults <- makeTwoYearsResults(eList, 1985, 2005)
}
\keyword{WRTDS}
\keyword{flow}

