% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_bootstrapping.R
\name{monthSeqToDec}
\alias{monthSeqToDec}
\title{monthSeqToDec}
\usage{
monthSeqToDec(monthSeq)
}
\arguments{
\item{monthSeq}{integer vector of months. Month 1 is considered Jan. 1850.}
}
\description{
Convert a sequence of month integers into their decimal years.
}
\examples{

months <- 1558:1600
monthSeqToDec(months)
}
