% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCIs.R
\name{plotConcHistBoot}
\alias{plotConcHistBoot}
\title{Graph of annual concentration, flow normalized concentration,
and confidence bands for flow normalized concentrations}
\usage{
plotConcHistBoot(eList, CIAnnualResults, plotFlowNorm = TRUE,
  col.pred = "green", printTitle = TRUE, cex.main = 1.1, ...)
}
\arguments{
\item{eList}{named list from EGRET package after running modelEstimation}

\item{CIAnnualResults}{data frame generated from ciBands (includes nBoot, probs, and blockLength attributes)}

\item{plotFlowNorm}{logical}

\item{col.pred}{character prediction color}

\item{printTitle}{logical}

\item{cex.main}{numeric title scale}

\item{\dots}{graphical parameters}
}
\description{
Uses the output of modelEstimation in the EGRET package (results in the named
list eList), and the data frame CIAnnualResults (produced by EGRETci package
using scripts described in the vignette) to produce a graph of annual
concentration, flow normalized concentration, and confidence bands for
flow-normalized concentrations.  In addition to the arguments listed below,
it will accept any additional arguments that are listed for the EGRET function
plotConcHist.
}
\examples{
library(EGRET)
eList <- Choptank_eList
nBoot <- 100
blockLength <- 200
\dontrun{
repAnnualResults <- vector(mode = "list", length = nBoot)
for(n in 1:nBoot){
   repAnnualResults[[n]] <- bootAnnual(eList, blockLength)
}

CIAnnualResults <- ciBands(eList, repAnnualResults)
plotConcHistBoot(eList, CIAnnualResults)
}
}

