% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCIs.R
\name{plotHistogramTrend}
\alias{plotHistogramTrend}
\title{plotHistogramTrend}
\usage{
plotHistogramTrend(eBoot, caseSetUp, xSeq = seq(-50, 50, 10), flux = TRUE,
  printTitle = TRUE, cex.main = 1.1, col.fill = "grey", ...)
}
\arguments{
\item{eBoot}{named list}

\item{caseSetUp}{data frame}

\item{xSeq}{vector}

\item{flux}{logical if TRUE, plots flux results, if FALSE plots concentration}

\item{printTitle}{logical if TRUE, includes title}

\item{cex.main}{numeric title font size}

\item{col.fill}{character fill color}
}
\description{
plotHistogramTrend
}
\examples{
library(EGRET)
eList <- Choptank_eList
\dontrun{
caseSetUp <- trendSetUp(eList)
eList <- setPA(eList)
eList <- setForBoot(eList)
eBoot <- wBT(eList,caseSetUp)
plotHistogramTrend(eBoot,flux=FALSE, xSeq = seq(-20,60,5))
plotHistogramTrend(eBoot,flux=TRUE, xSeq = seq(-20,60,5))
}
}

