% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCIs.R
\name{saveCB}
\alias{saveCB}
\title{saveCB}
\usage{
saveCB(eList)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}
}
\description{
saveCB
}
\examples{
library(EGRET)
eList <- Choptank_eList
\dontrun{
saveCB(eList)
}
}

