% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{setForBoot}
\alias{setForBoot}
\title{setForBoot}
\usage{
setForBoot(eList, windowY = 7, windowQ = 2, windowS = 0.5,
  edgeAdjust = TRUE)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes. Created from the EGRET package, after running \code{\link[EGRET]{modelEstimation}}.}

\item{windowY}{numeric}

\item{windowQ}{numeric}

\item{windowS}{numeric}

\item{edgeAdjust}{logical}
}
\value{
surfaces matrix
}
\description{
setForBoot
}
\examples{
library(EGRET)
eList <- Choptank_eList

bootSetUp <- setForBoot(eList)
}
\keyword{WRTDS}
\keyword{flow}

