% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGroupsBoot.R
\name{single_boot_run}
\alias{single_boot_run}
\title{Single Bootstrap Run}
\usage{
single_boot_run(iBoot, startSeed, eList, type_results, jitterOn, V,
  blockLength, type)
}
\arguments{
\item{iBoot}{description}

\item{startSeed}{setSeed value. Defaults to 494817. This is used to make repeatable output.}

\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{type_results}{data frame returned from either \code{\link[EGRET]{runGroups}}
or \code{\link[EGRET]{runPairs}} depending on context.}

\item{jitterOn}{logical, if TRUE, adds "jitter" to the data in an attempt to avoid some numerical problems.
Default = FALSE.  See Details below.}

\item{V}{numeric a multiplier for addition of jitter to the data, default = 0.2.}

\item{blockLength}{integer size of subset, expressed in days.  200 days has been found to be a good choice.}

\item{type}{Character either "pair" or "group".}
}
\value{
list of xConc, xFlux, pConc, pFlux
}
\description{
Single run of a bootstrap.
}
