% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{wBT}
\alias{wBT}
\title{Run EGRETci bootstrap}
\usage{
wBT(eList, caseSetUp, prob = c(0.025, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95,
  0.975), saveOutput = TRUE, fileName = "temp.txt")
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{caseSetUp}{data frame}

\item{prob}{vector of probabilities}

\item{saveOutput}{logical}

\item{fileName}{character}
}
\value{
eBoot
}
\description{
Run EGRETci bootstrap
}
\examples{
library(EGRET)
eList <- Choptank_eList
\dontrun{
caseSetUp <- trendSetUp(eList)
eList <- setPA(eList)
eList <- setForBoot(eList)
eBoot <- wBT(eList,caseSetUp)
}
}

