% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{wordLike}
\alias{wordLike}
\title{Assigns the narrative descriptors to the likelihood of change that is calculated by WBT}
\usage{
wordLike(likeList)
}
\arguments{
\item{likeList}{list}
}
\value{
character vector for [1] Upward trend in concentration,
[2] Downward trend in concentration, [3] Upward trend in flux,
[4] Downward trend in flux
}
\description{
Function called in \code{\link{wBT}} to convert numeric likelihood percentages to useful text.
}
\examples{
likeList <- c(0.01, 0.5, 0.55, 0.99)
Trends <- wordLike(likeList)
}

