% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{blockSample}
\alias{blockSample}
\title{blockSample}
\usage{
blockSample(localSample, blockLength)
}
\arguments{
\item{localSample}{Sample data frame}

\item{blockLength}{integer}
}
\value{
surfaces matrix
}
\description{
Get blockSample
}
\examples{
library(EGRET)
eList <- Choptank_eList
Sample <- eList$Sample
bsReturn <- blockSample(Sample, 25)
}
\keyword{WRTDS}
\keyword{flow}

