% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{blockSample}
\alias{blockSample}
\title{blockSample}
\usage{
blockSample(localSample, blockLength, startSeed = NA)
}
\arguments{
\item{localSample}{Sample data frame}

\item{blockLength}{integer size of subset.}

\item{startSeed}{setSeed value. Defaults to 494817. This is used to make repeatable output.}
}
\value{
surfaces matrix
}
\description{
Get a subset of the Sample data frame based on the user-specified blockLength.
}
\examples{
library(EGRET)
eList <- Choptank_eList
Sample <- eList$Sample
bsReturn <- blockSample(Sample, 25)
}
\keyword{WRTDS}
\keyword{flow}
