% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPairsBoot.R
\name{boot_message}
\alias{boot_message}
\title{Print message of bootstrapping results}
\usage{
boot_message(eList, type_results, bootOut, nBootGood, nBoot, type = "pair")
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{type_results}{List returned from either runPairs or runGroup.}

\item{bootOut}{List returned from either runPairs or runGroupBoot.}

\item{nBootGood}{Number of good results.}

\item{nBoot}{Number of requested bootstraps.}

\item{type}{Character can be "pair" or "group"}
}
\description{
Uses the results of either group or pair results and prints
of the standard message.
}
