% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCIs.R
\name{bootAnnual}
\alias{bootAnnual}
\title{bootAnnual}
\usage{
bootAnnual(eList, blockLength = 200)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{blockLength}{integer suggested value is 200}
}
\description{
bootAnnual One bootstrap run.
}
\examples{
library(EGRET)
eList <- Choptank_eList
\dontrun{
annualResults <- bootAnnual(eList)
}
}

