% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCIs.R
\name{ciBands}
\alias{ciBands}
\title{ciBands}
\usage{
ciBands(eList, repAnnualResults, probs = c(0.05, 0.95))
}
\arguments{
\item{eList}{named list}

\item{repAnnualResults}{named list returned from bootstrapping process}

\item{probs}{vector high and low confidence interval percentages}
}
\description{
ciBands
}
\examples{
library(EGRET)
eList <- Choptank_eList
nBoot <- 100
blockLength <- 200
\dontrun{

repAnnualResults <- vector(mode = "list", length = nBoot)
for(n = 1:nBoot){
   annualResults <- bootAnnual(eList, blockLength)
   repAnnualResults[[n]] <- bootAnnual(eList, blockLength)
}

CIAnnualResults <- ciBands(eList, repAnnualResults)

}
}

