% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCIs.R
\name{ciCalculations}
\alias{ciCalculations}
\title{ciCalculations}
\usage{
ciCalculations(eList, ...)
}
\arguments{
\item{eList}{named list}

\item{\dots}{optionally include nBoot, blockLength, or widthCI}
}
\description{
Interactive function to calculate WRTDS confidence bands
}
\examples{
library(EGRET)
eList <- Choptank_eList
\dontrun{
CIAnnualResults <- ciCalculations(eList)
}
}

