% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wBTCode.R
\docType{package}
\name{EGRETci-package}
\alias{EGRETci-package}
\title{EGRETci package for confidence interval analysis}
\description{
\tabular{ll}{
Package: \tab EGRETci\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
http://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
}
Collection of functions to evaluate uncertainty of results from water quality analysis using 
the Weighted Regressions on Time Discharge and Season (WRTDS) method. This package is an add-on 
to the EGRET package that performs the WRTDS analysis.
}
\author{
Robert M. Hirsch \email{rhirsch@usgs.gov}, Laura De Cicco \email{ldecicco@usgs.gov}
}
\references{
Hirsch, R.M., and De Cicco, L.A., 2014, User guide to Exploration and Graphics for RivEr Trends 
(EGRET) and dataRetrieval: R packages for hydrologic data: U.S. Geological Survey Techniques and Methods book 4, 
chap. A10, 94 p., \url{http://dx.doi.org/10.3133/tm4A10}

Hirsch, R.M., Archfield, S.A., and De Cicco, L.A., 2015, 
A bootstrap method for estimating uncertainty of water quality trends.  
Accepted for publication: Journal of Environmental Modelling and Software.
}
\keyword{graphics}
\keyword{statistics}
\keyword{streamflow}
\keyword{water-quality}

