% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{estSliceSurfacesSimpleAlt}
\alias{estSliceSurfacesSimpleAlt}
\title{surface slice}
\usage{
estSliceSurfacesSimpleAlt(eList, year)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes. Created from the EGRET package, 
after running either \code{\link[EGRET]{modelEstimation}} or \code{\link{setForBoot}}.}

\item{year}{integer year to perform WRTDS analysis}
}
\value{
surfaces matrix
}
\description{
Creates surface slice for one year.
}
\examples{
library(EGRET)
eList <- Choptank_eList
\dontrun{
caseSetUp <- trendSetUp(eList, nBoot=100, blockLength=200)
eList <- setForBoot(eList, caseSetUp)
surfaces <- estSliceSurfacesSimpleAlt(eList, 1990)
}
}
\keyword{WRTDS}
\keyword{flow}
