% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{makeCombo}
\alias{makeCombo}
\title{makeCombo}
\usage{
makeCombo(surfaces1, surfaces2)
}
\arguments{
\item{surfaces1}{vector returned from \code{\link{estSliceSurfacesSimpleAlt}}}

\item{surfaces2}{vector returned from \code{\link{estSliceSurfacesSimpleAlt}}}
}
\value{
surfaces matrix
}
\description{
Combine surface slices.
}
\examples{
surfaces1 <- c(1,2,3)
surfaces2 <- c(4, NA, 5)
surfaces <- makeCombo(surfaces1, surfaces2)
}
\keyword{WRTDS}
\keyword{flow}

