% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{makeTwoYearsResults}
\alias{makeTwoYearsResults}
\title{makeTwoYearsResults}
\usage{
makeTwoYearsResults(eList, year1, year2)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes. Created from the EGRET package, after running \code{\link[EGRET]{modelEstimation}}.}

\item{year1}{integer. Initial year of a 2-year trend comparison.}

\item{year2}{integer. Second year of a 2-year trend comparison.}
}
\value{
surfaces matrix
}
\description{
In bootstrap process computes the results for the initial year and second year.
}
\examples{
library(EGRET)
eList <- Choptank_eList

twoResultsWaterYear <- makeTwoYearsResults(eList, 1985, 2005)
}
\keyword{WRTDS}
\keyword{flow}

