% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{paVector}
\alias{paVector}
\title{paVector}
\usage{
paVector(year, paStart, paLong, vectorYear)
}
\arguments{
\item{year}{integer year}

\item{paStart}{integer starting month for period of analysis}

\item{paLong}{integer length of period of analysis}

\item{vectorYear}{numeric vector of years}
}
\value{
surfaces matrix
}
\description{
Creates paVector
}
\examples{
year <- 2000
paStart <- 10
paLong <- 12
vectorYear <- seq(1985,2005,5)
output <- paVector(year, paStart, paLong, vectorYear)
}
\keyword{WRTDS}
\keyword{flow}

