% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{paVector}
\alias{paVector}
\title{Create a paVector}
\usage{
paVector(year, paStart, paLong, vectorYear)
}
\arguments{
\item{year}{description}

\item{paStart}{description}

\item{paLong}{description}

\item{vectorYear}{description}
}
\description{
Internal doc for paVector
}
\keyword{internal}
