% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCIs.R
\name{plotConcHistBoot}
\alias{plotConcHistBoot}
\title{plotConcHistBoot}
\usage{
plotConcHistBoot(eList, CIAnnualResults, plotFlowNorm = TRUE,
  col.pred = "green", printTitle = TRUE, cex.main = 1.1, ...)
}
\arguments{
\item{eList}{named list}

\item{CIAnnualResults}{data frame from ciBands (needs nBoot, probs, and blockLength attributes)}

\item{plotFlowNorm}{logical}

\item{col.pred}{character prediction color}

\item{printTitle}{logical}

\item{cex.main}{numeric title scale}

\item{\dots}{graphical parameters}
}
\description{
plotConcHistBoot
}
\examples{
library(EGRET)
eList <- Choptank_eList
nBoot <- 100
blockLength <- 200
\dontrun{
repAnnualResults <- vector(mode = "list", length = nBoot)
for(n in 1:nBoot){
   repAnnualResults[[n]] <- bootAnnual(eList, blockLength)
}

CIAnnualResults <- ciBands(eList, repAnnualResults)
plotConcHistBoot(eList, CIAnnualResults)
}
}

