% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCIs.R
\name{plotConcHistBoot}
\alias{plotConcHistBoot}
\title{Graph of annual concentration, flow normalized concentration,
and confidence bands for flow normalized concentrations}
\usage{
plotConcHistBoot(eList, CIAnnualResults, yearStart = NA, yearEnd = NA,
  plotFlowNorm = TRUE, col.pred = "green", concMax = NA,
  printTitle = TRUE, cex.main = 1.1, ...)
}
\arguments{
\item{eList}{named list from EGRET package after running modelEstimation}

\item{CIAnnualResults}{data frame generated from ciBands (includes nBoot, probs, and blockLength attributes)}

\item{yearStart}{numeric is the calendar year containing the first estimated annual value to be plotted, default is NA (which allows it to be set automatically by the data)}

\item{yearEnd}{numeric is the calendar year just after the last estimated annual value to be plotted, default is NA (which allows it to be set automatically by the data)}

\item{plotFlowNorm}{logical variable if TRUE flow normalized line is plotted, if FALSE not plotted}

\item{col.pred}{character prediction color}

\item{concMax}{number specifying the maximum value to be used on the vertical axis, default is NA (which allows it to be set automatically by the data)}

\item{printTitle}{logical}

\item{cex.main}{numeric title scale}

\item{\dots}{graphical parameters}
}
\description{
Uses the output of modelEstimation in the EGRET package (results in the named
list eList), and the data frame CIAnnualResults (produced by EGRETci package
using scripts described in the vignette) to produce a graph of annual
concentration, flow normalized concentration, and confidence bands for
flow-normalized concentrations.  In addition to the arguments listed below,
it will accept any additional arguments that are listed for the EGRET function
plotConcHist.
}
\examples{
library(EGRET)
eList <- Choptank_eList
CIAnnualResults <- Choptank_CIAnnualResults
plotConcHistBoot(eList, CIAnnualResults)
plotConcHistBoot(eList, CIAnnualResults, yearStart=1990, yearEnd=2002)
\dontrun{
CIAnnualResults <- ciCalculations(eList, nBoot = 100, blockLength = 200)
plotConcHistBoot(eList, CIAnnualResults)
}
}

