% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotCIs.R
\name{plotFluxHistBoot}
\alias{plotFluxHistBoot}
\title{plotFluxHistBoot}
\usage{
plotFluxHistBoot(eList, CIAnnualResults, yearStart = NA, yearEnd = NA,
  plotFlowNorm = TRUE, fluxUnit = 9, col.pred = "green",
  printTitle = TRUE, cex.main = 1.1, ...)
}
\arguments{
\item{eList}{named list}

\item{CIAnnualResults}{data frame from ciBands (needs nBoot, probs, and blockLength attributes)}

\item{yearStart}{numeric is the calendar year containing the first estimated annual value to be plotted, default is NA (which allows it to be set automatically by the data)}

\item{yearEnd}{numeric is the calendar year just after the last estimated annual value to be plotted, default is NA (which allows it to be set automatically by the data)}

\item{plotFlowNorm}{logical variable if TRUE flow normalized line is plotted, if FALSE not plotted}

\item{fluxUnit}{number representing entry in pre-defined fluxUnit class array. \code{\link{printFluxUnitCheatSheet}}}

\item{col.pred}{character prediction color}

\item{printTitle}{logical}

\item{cex.main}{numeric title scale}

\item{\dots}{graphical parameters}
}
\description{
plotFluxHistBoot
}
\examples{
library(EGRET)
eList <- Choptank_eList
CIAnnualResults <- Choptank_CIAnnualResults
plotFluxHistBoot(eList, CIAnnualResults, fluxUnit=5)

\dontrun{
CIAnnualResults <- ciCalculations(eList, nBoot = 100, blockLength = 200)
plotFluxHistBoot(eList, CIAnnualResults, fluxUnit=5)
}
}

