% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{pVal}
\alias{pVal}
\title{pVal}
\usage{
pVal(s)
}
\arguments{
\item{s}{slope values from the bootstrap (already flipped)}
}
\description{
Computes the two-sided p value for the null hypothesis
}
\examples{
s <- c(0.01, 0.5, 0.55, 0.99)
pValue <- pVal(s)
}

