% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGroupBoot.R
\name{runGroupsBoot}
\alias{runGroupsBoot}
\title{runGroupsBoot}
\usage{
runGroupsBoot(eList, groupResults, nBoot = 100, startSeed = 494817,
  blockLength = 200)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{groupResults}{data frame returned from \code{EGRET::runGroups}}

\item{nBoot}{the maximum number of bootstrap replicates to be used, typically 100}

\item{startSeed}{setSeed value. Defaults to 494817. This is used to make repeatable output.}

\item{blockLength}{days, typically 200 is a good choice}
}
\description{
runGroupsBoot
}
\examples{
library(EGRET)
eList <- Choptank_eList

\dontrun{
groupResults <- runGroups(eList, 
                          group1firstYear = 1995, 
                          group1lastYear = 2004, 
                          group2firstYear = 2005, 
                          group2lastYear = 2014, 
                          windowSide = 7, wall = TRUE, 
                          sample1EndDate = "2004-10-30", 
                          paStart = 4, paLong = 2, 
                          verbose = FALSE)

boot_group_out <- runGroupsBoot(eList, groupResults)

plotHistogramTrend(eList, boot_group_out, caseSetUp=NA)
}
}
