% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPairsBoot.R
\name{runPairsBoot}
\alias{runPairsBoot}
\title{runPairsBoot}
\usage{
runPairsBoot(eList, pairResults, nBoot = 100, startSeed = 494817,
  blockLength = 200)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}

\item{pairResults}{data frame returned from \code{EGRET::runPairs}}

\item{nBoot}{the maximum number of bootstrap replicates to be used, typically 100}

\item{startSeed}{setSeed value. Defaults to 494817. This is used to make repeatable output.}

\item{blockLength}{days, typically 200 is a good choice}
}
\description{
runPairsBoot
}
\examples{
library(EGRET)
eList <- Choptank_Phos
year1 <- 1985
year2 <- 2014

\dontrun{
pairOut_2 <- runPairs(eList, year1, year2, windowSide = 7)

boot_pair_out <- runPairsBoot(eList, pairOut_2)
}
}
