% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wBTCode.R
\name{trendSetUp}
\alias{trendSetUp}
\title{Interactive setup for EGRETci}
\usage{
trendSetUp(eList)
}
\arguments{
\item{eList}{named list with at least the Daily, Sample, and INFO dataframes}
}
\value{
condition data frame with year1,yearData1,year2,yearData2,numSamples,nBoot,bootBreak,blockLength,confStop
}
\description{
Walks user through the set-up for a trend analysis
}
\examples{
library(EGRET)
eList <- Choptank_eList
\dontrun{
caseSetUp <- trendSetUp(eList)
}
}
\keyword{WRTDS}
\keyword{flow}

